fixed styling and strucure
