--- src/Signup.css
@@ -1,135 +1,0 @@
-* {
-    box-sizing: border-box;
-}
-
-.signup-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    height: auto;
-    background-color: #000;
-    color: white;
-    text-align: center;
-    max-width: 30rem;
-    width: 100%;
-    padding: 1.5rem;
-    border-radius: 0.5rem;
-    box-shadow: 0px 4px 10px rgba(255, 255, 255, 0.2);
-}
-
-.form-row {
-    display: flex;
-    gap: 0.5rem;
-    width: 100%;
-}
-
-.form-row input {
-    width: 50%;
-}
-
-.description {
-    font-family: 'Poppins', sans-serif;
-    font-weight: 400;
-    font-size: 1rem;
-    margin: 0.5rem 0 1.5rem 0;
-}
-
-form {
-    display: flex;
-    flex-direction: column;
-    gap: 0.8rem;
-    width: 100%;
-}
-
-input {
-    width: 100%;
-    padding: 0.7rem;
-    margin: 0.3rem 0;
-    border: none;
-    border-radius: 0.3rem;
-    background: #222;
-    color: white;
-    font-size: 0.9rem;
-    outline: none;
-}
-
-input[type="date"] {
-    color: rgba(255, 255, 255, 0.7);
-}
-
-input::placeholder {
-    color: rgba(255, 255, 255, 0.7);
-}
-
-button {
-    width: 100%;
-    padding: 0.7rem 1rem;
-    margin-top: 1rem;
-    background: white;
-    color: black;
-    border: none;
-    border-radius: 0.3rem;
-    font-size: 1rem;
-    cursor: pointer;
-    transition: background 0.3s;
-}
-
-button:hover {
-    background: #ddd;
-}
-
-.line-divider {
-    display: flex;
-    align-items: center;
-    text-align: center;
-    color: #999;
-    margin: 1rem 0;
-    width: 100%;
-}
-
-.line-divider::before,
-.line-divider::after {
-    content: "";
-    flex: 1;
-    border-top: 2px solid #555;
-    margin: 0 10px;
-}
-
-.google-login {
-    background: #db4437;
-    color: white;
-}
-
-.facebook-login {
-    background: #3b5998;
-    color: white;
-}
-
-.google-login:hover {
-    background: #c1351d;
-}
-
-.facebook-login:hover {
-    background: #2d4373;
-}
-
-.login-link {
-    margin-top: 1.5rem;
-    font-size: 0.9rem;
-}
-
-.login-link a {
-    color: #ff7b00;
-    text-decoration: none;
-}
-
-.login-link a:hover {
-    text-decoration: underline;
-}
-
-.error-message {
-    color: #ff5555;
-    font-size: 0.9rem;
-    margin-bottom: 1rem;
-}
