+++ src/pages/Landing.css
@@ -2,22 +2,26 @@
   box-sizing: border-box;
 }
 
+body{
+  background: url("../assets/bg_landing.jpg") center/cover no-repeat;
+  overflow-x: hidden;
+}
+
+.app-container{
+  height: 100%;
+}
+
 /* General Styles */
 .landing {
   text-align: center;
   color: white;
-  background: black;
   width: 100%;
   display: flex;
   flex-direction: column;
   align-items: center;
 }
 
-.logo {
-  margin: 1.25rem 0;
-  width: 6.25rem;
-  height: auto;
-}
+
 
 /* Hero Section */
 .hero {
@@ -28,7 +32,7 @@
   align-items: center;
   text-align: center;
   color: white;
-  padding: 0 1.25rem;
+  padding: 2rem ;
 }
 
 .hero-background {
@@ -37,7 +41,6 @@
   left: 0;
   width: 100%;
   height: 100%;
-  background: url("../assets/bg_landing.jpg") center/cover no-repeat;
   z-index: -1;
 }
 
@@ -49,6 +52,7 @@
   width: 100%;
   height: 100%;
   background: rgba(0, 0, 0, 0.5);
+  border-radius: 10px;
 }
 
 .hero-content {
@@ -94,7 +98,7 @@
 }
 
 .feature {
-  background: #222;
+  background: rgba(0, 0, 0, 0.8);
   padding: 1.875rem;
   border-radius: 0.625rem;
   width: 18.75rem;
@@ -118,10 +122,11 @@
 
 /* Call-to-Action Section */
 .cta-section {
-  margin-top: 3.125rem;
-  background: #111;
+  margin: 3.125rem;
+  background: rgba(0, 0, 0, 0.8);
   padding: 3.125rem;
-  width: 100%;
+  border-radius: 10px;
+
 }
 
 .cta-section h2 {
@@ -129,16 +134,3 @@
   margin-bottom: 1.25rem;
 }
 
-/* Footer */
-.footer {
-  background: #000;
-  padding: 1.25rem;
-  width: 100%;
-  text-align: center;
-  margin-top: auto;
-}
-
-.footer p {
-  margin: 0;
-  font-size: 0.9rem;
-}
