+++ src/pages/Login.css
@@ -35,3 +35,24 @@
 .forgot-password:hover {
   color: #fff;
 }
+
+.login-container{
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  height: auto;
+  background: rgba(0, 0, 0, 0.8);
+  color: white;
+  text-align: center;
+  max-width: 30rem;
+  width: 100%;
+  padding: 2.5rem;
+  border-radius: 0.5rem;
+  box-shadow: 0px 4px 10px rgba(255, 255, 255, 0.2);
+}
+
+body{
+  height: 100%;
+
+}
